pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "anny"
version = "0.1.0"
authors = [
    { name = "Romain Brégier", email = "romain.bregier@naverlabs.com" }
]
description="A parametric body model based on MakeHuman for PyTorch."
readme = "README.md"
license-files = ["LICENSE", "LICENSE_THINGS"]

dependencies = [
    "torch>=2.0.0",
    "roma>=1.5.2.1",
    "numpy"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"anny" = ["data/*"]

[project.optional-dependencies]
examples = [
    "gradio",
    "jsonargparse",
    "trimesh"
]
warp = [
    "warp-lang"
]
frontend = [
    "gradio>=4.0.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "jsonargparse>=4.0.0",
    "trimesh>=4.0.0"
]
vision = [
    "mediapipe>=0.10.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0"
]
dev = [
    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "pytest-xdist>=3.0.0",  # Parallel testing
    "hypothesis>=6.0.0",  # Property-based testing

    # Code Quality
    "black>=23.0.0",
    "isort>=5.13.0",
    "flake8>=6.0.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",

    # Security
    "bandit[toml]>=1.7.0",
    "safety>=2.0.0",

    # Pre-commit
    "pre-commit>=3.0.0",

    # Documentation
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",

    # Type stubs
    "types-requests>=2.0.0",
    "types-pillow>=10.0.0",

    # Profiling & Debugging
    "pytest-profiling>=1.7.0",
    "memory-profiler>=0.61.0",
    "line-profiler>=4.0.0",

    # Build tools
    "build>=1.0.0",
    "twine>=4.0.0"
]
# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
split_on_trailing_comma = true
skip_gitignore = true
known_first_party = ["src", "anny"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "security: marks tests as security tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/migrations/*",
    "*/examples/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "roma.*",
    "trimesh.*",
    "gradio.*",
    "mediapipe.*",
    "cv2.*",
]
ignore_missing_imports = true

# Ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit
    "T",   # flake8-print
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex (we'll handle this manually)
    "S101",  # use of assert (needed for tests)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"tests/**/*.py" = ["S101", "S105", "S106"]  # assert, hardcoded passwords in tests

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests", "test", "benchmarks"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*/tests/*.py"]
